+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Next
+
+- feat: add `$app_install_source` property ([#442](https://github.com/PostHog/posthog-ios/pull/442))
+
 - chore: support new surveys color options for ios+flutter ([#440](https://github.com/PostHog/posthog-ios/pull/440))
 
 - feat: support thumbs up/down surveys for ios ([#437](https://github.com/PostHog/posthog-ios/pull/437))